Move encoder core
